requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.20
numpy==1.24.3
matplotlib==3.7.2
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3

